# openapi.yaml
openapi: 3.1.0
info:
  title: RGB Lightning Node
  description: |-
    This is the OpenAPI specification for the
    [RGB Lightning Node](https://github.com/RGB-Tools/rgb-lightning-node) APIs.

    On mainnet, Lightning APIs return HTTP 403 with name
    LightningUnsupportedOnMainnet and the message: "RLN on mainnet currently
    supports only on-chain methods. Lightning APIs are not supported."
    This includes channel/peer operations, Lightning invoices and decoding,
    payments, async payments, swaps and swap-string decoding, and onion messages.
    The configured node network controls the restriction, including while locked.
    Authentication and request extraction still apply before handler execution.
    Bitcoin/RGB on-chain APIs (including RGB invoices, transfers and asset linking)
    and shared administration/identity APIs retain their existing requirements.
    Mainnet unlock starts the wallet and signer without a Lightning runtime.
    Historical Lightning records remain inactive and do not prevent unlock. They
    are not decoded, replayed or recovered. The supported mainnet rollout assumes
    no unresolved historical Lightning obligations; unlock does not verify that
    assumption or monitor old channels. Lightning APIs on supported non-mainnet
    networks are unchanged.
  license:
    name: MIT
    url: https://mit-license.org/
  version: 0.1.0
servers:
  - url: http://localhost:3001
  - url: http://localhost:3002
  - url: http://localhost:3003
...
paths:
...
  /apay/new:
    post:
      tags:
        - Async Payments
      summary: Send a new order with async-payment hashes
      description: Derive deterministic payment hashes for async payments from the unlocked wallet. Send an async_order.new JSON-RPC request to the connected invoice-host peer so it can track those hashes and process asynchronous payments. A peer connection alone is not sufficient; the recipient-client must already have a live channel with the invoice-host peer. 
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsyncOrderNewRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Host accepted the async order request and returned order state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOrderNewResponse'
  /apay/outboundinvoice:
    post:
      tags:
        - Async Payments
      summary: Request an outbound LN invoice
      description: Request an outbound invoice from the connected recipient-client peer for an async payment claim session. A peer connection alone is not enough; the invoice-host must already have a live channel with the recipient-client peer.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsyncOrderOutboundInvoiceRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Recipient peer returned an outbound invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOrderOutboundInvoiceResponse'
  /assetbalance:
    post:
      tags:
        - RGB
      summary: Get the balance of an asset
      description: Get the balance for the provided RGB asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetBalanceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetBalanceResponse'
  /assetlink:
    post:
      tags:
        - RGB
      summary: Create a contract link
      description: Link a parent IFA contract to a child IFA contract using a link-right UTXO
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetLinkRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetLinkResponse'
  /assetmetadata:
    post:
      tags:
        - RGB
      summary: Get the metadata of an asset
      description: Get the metadata for the provided RGB asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetMetadataRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetMetadataResponse'
  /backup:
    post:
      tags:
        - Other
      summary: Backup the node
      description: Create a backup of the node's data
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /btcbalance:
    post:
      tags:
        - On-chain
      summary: Get the BTC balance
      description: Get the node's bitcoin balance for the vanilla and colored wallets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BtcBalanceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BtcBalanceResponse'
  /cancelhodlinvoice:
    post:
      tags:
        - Invoices
      summary: Cancel a HODL invoice
      description: Cancel a held HTLC for a HODL invoice. Rejects cancellation if a settlement is already in progress.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCancelRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /changepassword:
    post:
      tags:
        - Other
      summary: Change the password
      description: Change the node's password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /checkindexerurl:
    post:
      tags:
        - Other
      summary: Check an indexer URL
      description: Check the given indexer URL is valid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckIndexerUrlRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckIndexerUrlResponse'
  /checkproxyendpoint:
    post:
      tags:
        - Other
      summary: Check a proxy endpoint
      description: Check the given proxy endpoint is valid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckProxyEndpointRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /claimhodlinvoice:
    post:
      tags:
        - Invoices
      summary: Claim a HODL invoice
      description: Claim a held HTLC for a HODL invoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimHodlInvoiceRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimHodlInvoiceResponse'
  /closechannel:
    post:
      tags:
        - Channels
      summary: Close a channel
      description: Close a LN channel cooperatively or forcibly
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseChannelRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /connectpeer:
    post:
      tags:
        - Peers
      summary: Connect to a peer
      description: Connect to the provided LN peer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectPeerRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /createutxos:
    post:
      tags:
        - RGB
      summary: Create UTXOs
      description: Create UTXOs to be used for RGB operations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUtxosRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /decodelninvoice:
    post:
      tags:
        - Invoices
      summary: Decode a LN invoice
      description: Decode the provided LN invoice string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecodeLNInvoiceRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeLNInvoiceResponse'
  /decodergbinvoice:
    post:
      tags:
        - RGB
      summary: Decode an RGB invoice
      description: Decode the provided RGB invoice string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecodeRGBInvoiceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeRGBInvoiceResponse'
  /decodeswapstring:
    post:
      tags:
        - Swaps
      summary: Decode a swapstring
      description: Decode the provided swapstring
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecodeSwapstringRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeSwapstringResponse'
  /disconnectpeer:
    post:
      tags:
        - Peers
      summary: Disconnect from a peer
      description: Disconnect from the provided LN peer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisconnectPeerRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /estimatefee:
    post:
      tags:
        - On-chain
      summary: Get fee estimation
      description: Get on-chain fee estimation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstimateFeeRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateFeeResponse'
  /failtransfers:
    post:
      tags:
        - RGB
      summary: Fail RGB transfers
      description: Set the status for eligible RGB transfers to
        `TransferStatus::Failed`.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailTransfersRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailTransfersResponse'
...
  /rgbsweeps/recover:
    post:
      tags: [RGB]
      summary: Recover a quarantined RGB channel sweep
      description: >-
        Archives the receipt before recovery. Resume retries the original preparation after
        restoring its batch/files. Reprepare requires a Failed batch, durable evidence that
        signing never started, no cached signed transaction and unbroadcast LDK inputs.
        Resolve requires an already consumed batch and a confirmed transaction and keeps
        the inputs blocked from new preparation. Requires an unlocked Lightning node and
        admin or explicit endpoint permission. Corrupt receipts must be restored from backup.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RgbSweepRecoveryRequest'
      responses:
        '200':
          description: Recovery completed; the original receipt remains archived
        '400':
          description: Stale record_id, unsafe action or unavailable recovery prerequisite
  /getassetmedia:
    post:
      tags:
        - RGB
      summary: Get an asset media
      description: Get the hex string of the media bytes of the provided media digest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetAssetMediaRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAssetMediaResponse'
  /getchannelid:
    post:
      tags:
        - Channels
      summary: Get a channel's ID
      description: Get a channel's ID from its former temporary channel ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetChannelIdRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetChannelIdResponse'
  /getconsignment:
    post:
      tags:
        - RGB
      summary: Get a send consignment
      description: Get the hex string of the consignment bytes for a send transfer,
        identified by its asset ID and txid (it also covers burns, where the
        consignment is the proof of the burn). Useful to hand the consignment to
        the counterparty out-of-band when the node filesystem is not accessible
        from the API caller.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetConsignmentRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetConsignmentResponse'
  /getpayment:
    post:
      tags:
        - Payments
      summary: Get a payment
      description: Get a payment by its payment hash
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPaymentRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPaymentResponse'
  /getswap:
    post:
      tags:
        - Swaps
      summary: Get a swap
      description: Get a swap by its payment hash
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetSwapRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSwapResponse'
  /burn:
    post:
      tags:
        - RGB
      summary: Burn RGB assets
      description: Burn RGB assets on-chain. For BFA assets a 32-byte burn_recipient
        (hex) is required, telling where the burned amount is released. The burn
        consignment, which proves the burn, can then be retrieved with
        /getconsignment using the asset ID and the returned txid.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BurnRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BurnResponse'
  /inflate:
    post:
      tags:
        - RGB
      summary: Inflate RGB assets
      description: Inflate RGB assets on-chain
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InflateRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InflateResponse'
  /init:
    post:
      tags:
        - Other
      summary: Init the node
      description: Initialize a new node, optionally providing a mnemonic.
        If no mnemonic is provided a new one is generated.
        Note that a mnemonic alone is not sufficient to recover RGB state, which also requires client-side data.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitResponse'
  /invoicestatus:
    post:
      tags:
        - Invoices
      summary: Get an invoice status
      description: Get the status of the provided LN invoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceStatusRequest'
      responses:
        '403':
          $ref: '#/components/responses/LightningUnsupportedOnMainnet'
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceStatusResponse'
  /issueassetcfa:
    post:
      tags:
        - RGB
      summary: Issue an RGB CFA asset
      description: Issue an RGB CFA asset. To provide a media first call the /postassetmedia API.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAssetCFARequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAssetCFAResponse'
  /issueassetifa:
    post:
      tags:
        - RGB
      summary: Issue an RGB IFA asset
      description: Issue an RGB IFA asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAssetIFARequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAssetIFAResponse'
...
components:
...
  schemas:
...
    BtcBalanceRequest:
      type: object
      required:
        - skip_sync
      properties:
        skip_sync:
          type: boolean
          example: false
...
    EmptyResponse:
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Biscuit
security:
  - bearerAuth: []